Moved MapLayer model from service-base to service-admin
