+++ service-admin/pom.xml
@@ -23,6 +23,11 @@
             <groupId>fi.nls.oskari.service</groupId>
             <artifactId>oskari-users</artifactId>
         </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
