+++ service-admin/src/test/java/org/oskari/admin/MapLayerTest.java
@@ -1,4 +1,4 @@
-package org.oskari.data.model;
+package org.oskari.admin;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.DeserializationFeature;
