+++ service-admin/src/test/resources/org/oskari/admin/MapLayer.json
